+++ app/controllers/admin/comics_controller.rb
