+++ app/controllers/admin/magazine_volumes_controller.rb
