+++ app/controllers/admin/magazines_controller.rb
