+++ app/views/admin/comic_chapters/index.html.erb
@@ -20,9 +20,9 @@
         <td><%= comic_chapter.number %></td>
         <td><%= comic_chapter.subtitle %></td>
         <td><%= comic_chapter.total_point %></td>
-        <td><%= link_to 'Show', comic_chapter %></td>
-        <td><%= link_to 'Edit', edit_comic_chapter_path(comic_chapter) %></td>
-        <td><%= link_to 'Destroy', comic_chapter, method: :delete, data: { confirm: 'Are you sure?' } %></td>
+        <td><%= link_to 'Show', admin_comic_chapters_path %></td>
+        <td><%= link_to 'Edit', edit_admin_comic_chapter_path(comic_chapter) %></td>
+        <td><%= link_to 'Destroy', admin_comic_chapter_path(comic_chapter), method: :delete, data: { confirm: 'Are you sure?' } %></td>
       </tr>
     <% end %>
   </tbody>
@@ -30,4 +30,4 @@
 
 <br>
 
-<%= link_to 'New Comic chapter', new_comic_chapter_path %>
+<%= link_to 'New Comic chapter', new_admin_comic_chapter_path %>
