+++ app/views/admin/comic_chapters/index.json.jbuilder
