+++ app/views/admin/comic_chapters/show.html.erb
