+++ app/views/admin/comic_chapters/show.json.jbuilder
