+++ app/views/admin/comics/edit.html.erb
