+++ app/views/admin/comics/index.html.erb
