+++ app/views/admin/comics/index.json.jbuilder
