+++ app/views/admin/comics/new.html.erb
