+++ app/views/admin/comics/show.html.erb
