+++ app/views/admin/comics/show.json.jbuilder
