+++ app/views/admin/magazine_volumes/_form.html.erb
