+++ app/views/admin/magazine_volumes/index.html.erb
