+++ app/views/admin/magazine_volumes/index.json.jbuilder
