+++ app/views/admin/magazine_volumes/show.html.erb
