+++ app/views/admin/magazine_volumes/show.json.jbuilder
