+++ app/views/admin/magazines/_form.html.erb
