+++ app/views/admin/magazines/edit.html.erb
