+++ app/views/admin/magazines/index.html.erb
