+++ app/views/admin/magazines/index.json.jbuilder
