+++ app/views/admin/magazines/show.json.jbuilder
